+++ src/main/java/blackjack/domain/card/CardShuffleMachine.java
@@ -1,11 +1,11 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class CardShuffleMachine implements CardPickMachine{
+public class CardShuffleMachine implements CardPickMachine {
 
     private static final List<Integer> cardIndex;
     private static final int CARD_NUMBER = Suit.values().length * Symbol.values().length;
