+++ src/main/java/blackjack/domain/player/AbstractPlayer.java
@@ -1,4 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.player;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
+import blackjack.domain.result.Match;
 
 public abstract class AbstractPlayer implements Player {
