+++ src/main/java/blackjack/domain/player/Dealer.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.player;
+
+import blackjack.domain.card.Deck;
+import blackjack.domain.result.Match;
 
 public class Dealer extends AbstractPlayer implements Player {
