+++ src/main/java/blackjack/domain/result/MatchResult.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
