+++ src/test/java/blackjack/domain/card/AlwaysAscNumberMachine.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.List;
 import java.util.stream.Collectors;
