+++ src/test/java/blackjack/domain/result/MatchTest.java
@@ -1,10 +1,17 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Suit;
+import blackjack.domain.card.Symbol;
+import blackjack.domain.player.Dealer;
+import blackjack.domain.player.Guest;
+import blackjack.domain.player.Player;
 
 class MatchTest {
